Incluindo odometro e litros na requisição.

+++ src/main/java/core/entity/PagamentoCombustivel.kt
@@ -1,5 +1,6 @@
 package core.entity
 
+import java.math.BigDecimal
 import javax.persistence.Entity
 import javax.persistence.JoinColumn
 import javax.persistence.ManyToOne
@@ -17,5 +18,5 @@ open class PagamentoCombustivel : Pagamento<TipoDespesaCombustivel>() {
 
     var odometro: Int? = null
 
-    var litros: Int? = null
+    var litros: BigDecimal? = null
 }
